tools: Fix exit code in project-version.sh (#19106)
Piping to the while doesn't work, as it results in the while being
executed in a subshell and so the main shell doesn't get EXIT=1. Use
process substitution instead.

Then apply a version bump to changelogger that should have been in
#19102 but was missed because this error made it not fail CI.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/637129784